Rename QnA Advisor Message model, related
